migrate tests to flutter_command_test
