+++ packages/flutter_tools/test/commands.shard/permeable/build_bundle_test.dart
@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:typed_data';
-
 import 'package:args/command_runner.dart';
 import 'package:file/memory.dart';
 import 'package:flutter_tools/src/base/file_system.dart';
@@ -14,7 +12,6 @@ import 'package:flutter_tools/src/bundle.dart';
 import 'package:flutter_tools/src/bundle_builder.dart';
 import 'package:flutter_tools/src/cache.dart';
 import 'package:flutter_tools/src/commands/build_bundle.dart';
-import 'package:flutter_tools/src/convert.dart';
 import 'package:flutter_tools/src/features.dart';
 import 'package:flutter_tools/src/globals.dart' as globals;
 import 'package:flutter_tools/src/project.dart';
@@ -506,343 +503,6 @@ void main() {
     FileSystem: fsFactory,
     ProcessManager: () => FakeProcessManager.any(),
   });
-
-  testUsingContext('--dart-define-from-file successfully forwards values to build env', () async {
-    globals.fs.file(globals.fs.path.join('lib', 'main.dart')).createSync(recursive: true);
-    globals.fs.file('pubspec.yaml').createSync();
-    globals.fs.file('.packages').createSync();
-    await globals.fs.file('config1.json').writeAsString(
-      '''
-        {
-          "kInt": 1,
-          "kDouble": 1.1,
-          "name": "denghaizhu",
-          "title": "this is title from config json file",
-          "nullValue": null,
-          "containEqual": "sfadsfv=432f"
-        }
-      '''
-    );
-    await globals.fs.file('config2.json').writeAsString(
-        '''
-        {
-          "body": "this is body from config json file"
-        }
-      '''
-    );
-    final CommandRunner<void> runner = createTestCommandRunner(BuildBundleCommand(
-      logger: BufferLogger.test(),
-    ));
-
-    await runner.run(<String>[
-      'bundle',
-      '--no-pub',
-      '--dart-define-from-file=config1.json',
-      '--dart-define-from-file=config2.json',
-    ]);
-  }, overrides: <Type, Generator>{
-    BuildSystem: () => TestBuildSystem.all(BuildResult(success: true), (Target target, Environment environment) {
-      expect(
-        _decodeDartDefines(environment),
-        containsAllInOrder(const <String>[
-          'kInt=1',
-          'kDouble=1.1',
-          'name=denghaizhu',
-          'title=this is title from config json file',
-          'nullValue=null',
-          'containEqual=sfadsfv=432f',
-          'body=this is body from config json file',
-        ]),
-      );
-    }),
-    FileSystem: fsFactory,
-    ProcessManager: () => FakeProcessManager.any(),
-  });
-
-  testUsingContext('values from --dart-define supersede values from --dart-define-from-file', () async {
-    globals.fs
-        .file(globals.fs.path.join('lib', 'main.dart'))
-        .createSync(recursive: true);
-    globals.fs.file('pubspec.yaml').createSync();
-    globals.fs.file('.packages').createSync();
-    globals.fs.file('.env').writeAsStringSync('''
-        MY_VALUE=VALUE_FROM_ENV_FILE
-      ''');
-    final CommandRunner<void> runner =
-        createTestCommandRunner(BuildBundleCommand(
-      logger: BufferLogger.test(),
-    ));
-
-    await runner.run(<String>[
-      'bundle',
-      '--no-pub',
-      '--dart-define=MY_VALUE=VALUE_FROM_COMMAND',
-      '--dart-define-from-file=.env',
-    ]);
-
-  }, overrides: <Type, Generator>{
-    BuildSystem: () => TestBuildSystem.all(BuildResult(success: true),
-            (Target target, Environment environment) {
-          expect(
-            _decodeDartDefines(environment),
-            containsAllInOrder(const <String>[
-              'MY_VALUE=VALUE_FROM_ENV_FILE',
-              'MY_VALUE=VALUE_FROM_COMMAND',
-            ]),
-          );
-        }),
-    FileSystem: fsFactory,
-    ProcessManager: () => FakeProcessManager.any(),
-  });
-
-  testUsingContext('--dart-define-from-file correctly parses a valid env file', () async {
-    globals.fs
-        .file(globals.fs.path.join('lib', 'main.dart'))
-        .createSync(recursive: true);
-    globals.fs.file('pubspec.yaml').createSync();
-    globals.fs.file('.packages').createSync();
-    await globals.fs.file('.env').writeAsString('''
-        # comment
-        kInt=1
-        kDouble=1.1 # should be double
-
-        name=piotrfleury
-        title=this is title from config env file
-        empty=
-
-        doubleQuotes="double quotes 'value'#=" # double quotes
-        singleQuotes='single quotes "value"#=' # single quotes
-        backQuotes=`back quotes "value" '#=` # back quotes
-
-        hashString="some-#-hash-string-value"
-
-        # Play around with spaces around the equals sign.
-        spaceBeforeEqual =value
-        spaceAroundEqual = value
-        spaceAfterEqual= value
-
-      ''');
-    await globals.fs.file('.env2').writeAsString('''
-        # second comment
-
-        body=this is body from config env file
-      ''');
-    final CommandRunner<void> runner =
-        createTestCommandRunner(BuildBundleCommand(
-      logger: BufferLogger.test(),
-    ));
-
-    await runner.run(<String>[
-      'bundle',
-      '--no-pub',
-      '--dart-define-from-file=.env',
-      '--dart-define-from-file=.env2',
-    ]);
-  }, overrides: <Type, Generator>{
-    BuildSystem: () => TestBuildSystem.all(BuildResult(success: true),
-            (Target target, Environment environment) {
-          expect(
-            _decodeDartDefines(environment),
-            containsAllInOrder(const <String>[
-              'kInt=1',
-              'kDouble=1.1',
-              'name=piotrfleury',
-              'title=this is title from config env file',
-              'empty=',
-              "doubleQuotes=double quotes 'value'#=",
-              'singleQuotes=single quotes "value"#=',
-              'backQuotes=back quotes "value" \'#=',
-              'hashString=some-#-hash-string-value',
-              'spaceBeforeEqual=value',
-              'spaceAroundEqual=value',
-              'spaceAfterEqual=value',
-              'body=this is body from config env file'
-            ]),
-          );
-        }),
-    FileSystem: fsFactory,
-    ProcessManager: () => FakeProcessManager.any(),
-  });
-
-  testUsingContext('--dart-define-from-file option env file throws a ToolExit when .env file contains a multiline value', () async {
-    globals.fs
-        .file(globals.fs.path.join('lib', 'main.dart'))
-        .createSync(recursive: true);
-    globals.fs.file('pubspec.yaml').createSync();
-    globals.fs.file('.packages').createSync();
-    await globals.fs.file('.env').writeAsString('''
-        # single line value
-        name=piotrfleury
-
-        # multi-line value
-        multiline = """ Welcome to .env demo
-        a simple counter app with .env file support
-        for more info, check out the README.md file
-        Thanks! """ # This is the welcome message that will be displayed on the counter app
-
-      ''');
-    final CommandRunner<void> runner =
-        createTestCommandRunner(BuildBundleCommand(
-      logger: BufferLogger.test(),
-    ));
-
-    expect(() => runner.run(<String>[
-      'bundle',
-      '--no-pub',
-      '--dart-define-from-file=.env',
-    ]), throwsToolExit(message: 'Multi-line value is not supported: multiline = """ Welcome to .env demo'));
-  }, overrides: <Type, Generator>{
-    BuildSystem: () => TestBuildSystem.all(BuildResult(success: true)),
-    FileSystem: fsFactory,
-    ProcessManager: () => FakeProcessManager.any(),
-  });
-
-  testUsingContext('--dart-define-from-file option works with mixed file formats',
-      () async {
-    globals.fs
-        .file(globals.fs.path.join('lib', 'main.dart'))
-        .createSync(recursive: true);
-    globals.fs.file('pubspec.yaml').createSync();
-    globals.fs.file('.packages').createSync();
-    await globals.fs.file('.env').writeAsString('''
-        kInt=1
-        kDouble=1.1
-        name=piotrfleury
-        title=this is title from config env file
-      ''');
-    await globals.fs.file('config.json').writeAsString('''
-        {
-          "body": "this is body from config json file"
-        }
-      ''');
-    final CommandRunner<void> runner =
-        createTestCommandRunner(BuildBundleCommand(
-      logger: BufferLogger.test(),
-    ));
-
-    await runner.run(<String>[
-      'bundle',
-      '--no-pub',
-      '--dart-define-from-file=.env',
-      '--dart-define-from-file=config.json',
-    ]);
-  }, overrides: <Type, Generator>{
-    BuildSystem: () => TestBuildSystem.all(BuildResult(success: true),
-            (Target target, Environment environment) {
-          expect(
-            _decodeDartDefines(environment),
-            containsAllInOrder(const <String>[
-              'kInt=1',
-              'kDouble=1.1',
-              'name=piotrfleury',
-              'title=this is title from config env file',
-              'body=this is body from config json file',
-            ]),
-          );
-        }),
-    FileSystem: fsFactory,
-    ProcessManager: () => FakeProcessManager.any(),
-  });
-
-  testUsingContext('test --dart-define-from-file option if conflict', () async {
-    globals.fs.file(globals.fs.path.join('lib', 'main.dart')).createSync(recursive: true);
-    globals.fs.file('pubspec.yaml').createSync();
-    globals.fs.file('.packages').createSync();
-    await globals.fs.file('config1.json').writeAsString(
-        '''
-        {
-          "kInt": 1,
-          "kDouble": 1.1,
-          "name": "denghaizhu",
-          "title": "this is title from config json file"
-        }
-      '''
-    );
-    await globals.fs.file('config2.json').writeAsString(
-        '''
-        {
-          "kInt": "2"
-        }
-      '''
-    );
-    final CommandRunner<void> runner = createTestCommandRunner(BuildBundleCommand(
-      logger: BufferLogger.test(),
-    ));
-
-    await runner.run(<String>[
-      'bundle',
-      '--no-pub',
-      '--dart-define-from-file=config1.json',
-      '--dart-define-from-file=config2.json',
-    ]);
-  }, overrides: <Type, Generator>{
-    BuildSystem: () => TestBuildSystem.all(BuildResult(success: true), (Target target, Environment environment) {
-      expect(
-        _decodeDartDefines(environment),
-        containsAllInOrder(<String>['kInt=2', 'kDouble=1.1', 'name=denghaizhu', 'title=this is title from config json file']),
-      );
-    }),
-    FileSystem: fsFactory,
-    ProcessManager: () => FakeProcessManager.any(),
-  });
-
-  testUsingContext('test --dart-define-from-file option by invalid file type', () {
-    globals.fs.file(globals.fs.path.join('lib', 'main.dart')).createSync(recursive: true);
-    globals.fs.file('pubspec.yaml').createSync();
-    globals.fs.file('.packages').createSync();
-    globals.fs.directory('config').createSync();
-    final CommandRunner<void> runner = createTestCommandRunner(BuildBundleCommand(
-      logger: BufferLogger.test(),
-    ));
-
-    expect(() => runner.run(<String>[
-      'bundle',
-      '--no-pub',
-      '--dart-define-from-file=config',
-    ]), throwsToolExit(message: 'Did not find the file passed to "--dart-define-from-file". Path: config'));
-  }, overrides: <Type, Generator>{
-    FileSystem: fsFactory,
-    BuildSystem: () => TestBuildSystem.all(BuildResult(success: true)),
-    ProcessManager: () => FakeProcessManager.any(),
-  });
-
-  testUsingContext('test --dart-define-from-file option by corrupted json', () async {
-    globals.fs.file(globals.fs.path.join('lib', 'main.dart')).createSync(recursive: true);
-    globals.fs.file('pubspec.yaml').createSync();
-    globals.fs.file('.packages').createSync();
-    await globals.fs.file('config.json').writeAsString(
-        '''
-        {
-          "kInt": 1Error json format
-          "kDouble": 1.1,
-          "name": "denghaizhu",
-          "title": "this is title from config json file"
-        }
-      '''
-    );
-    final CommandRunner<void> runner = createTestCommandRunner(BuildBundleCommand(
-      logger: BufferLogger.test(),
-    ));
-
-    expect(() => runner.run(<String>[
-      'bundle',
-      '--no-pub',
-      '--dart-define-from-file=config.json',
-    ]), throwsToolExit(message: 'Json config define file "--dart-define-from-file=config.json" format err'));
-  }, overrides: <Type, Generator>{
-    FileSystem: fsFactory,
-    BuildSystem: () => TestBuildSystem.all(BuildResult(success: true)),
-    ProcessManager: () => FakeProcessManager.any(),
-  });
-}
-
-Iterable<String> _decodeDartDefines(Environment environment) {
-  final String encodedDefines = environment.defines[kDartDefines]!;
-  const Utf8Decoder byteDecoder = Utf8Decoder();
-  return encodedDefines
-      .split(',')
-      .map<Uint8List>(base64.decode)
-      .map<String>(byteDecoder.convert);
 }
 
 class FakeBundleBuilder extends Fake implements BundleBuilder {
