+++ packages/flutter_tools/test/general.shard/runner/flutter_command_test.dart
@@ -4,6 +4,7 @@
 
 import 'dart:async';
 import 'dart:io' as io;
+import 'dart:typed_data';
 
 import 'package:args/command_runner.dart';
 import 'package:file/memory.dart';
@@ -17,8 +18,10 @@ import 'package:flutter_tools/src/base/signals.dart';
 import 'package:flutter_tools/src/base/time.dart';
 import 'package:flutter_tools/src/base/user_messages.dart';
 import 'package:flutter_tools/src/build_info.dart';
+import 'package:flutter_tools/src/build_system/build_system.dart';
 import 'package:flutter_tools/src/cache.dart';
 import 'package:flutter_tools/src/commands/run.dart';
+import 'package:flutter_tools/src/convert.dart';
 import 'package:flutter_tools/src/dart/pub.dart';
 import 'package:flutter_tools/src/device.dart';
 import 'package:flutter_tools/src/globals.dart' as globals;
@@ -41,6 +44,9 @@ void main() {
     late FakeClock clock;
     late FakeProcessInfo processInfo;
     late MemoryFileSystem fileSystem;
+    late Platform platform;
+    late FileSystemUtils fileSystemUtils;
+    late Logger logger;
     late FakeProcessManager processManager;
     late PreRunValidator preRunValidator;
 
@@ -56,6 +62,9 @@ void main() {
       processInfo = FakeProcessInfo();
       processInfo.maxRss = 10;
       fileSystem = MemoryFileSystem.test();
+      platform = FakePlatform();
+      fileSystemUtils = FileSystemUtils(fileSystem: fileSystem, platform: platform);
+      logger = BufferLogger.test();
       processManager = FakeProcessManager.empty();
       preRunValidator = PreRunValidator(fileSystem: fileSystem);
     });
@@ -704,6 +713,364 @@ void main() {
       });
     });
 
+    group('--dart-define-from-file', () {
+
+      late FlutterCommand dummyCommand;
+      late CommandRunner<void> dummyCommandRunner;
+
+      setUp(() {
+        dummyCommand = DummyFlutterCommand()..usesDartDefineOption();
+        dummyCommandRunner = createTestCommandRunner(dummyCommand);
+      });
+
+      testUsingContext('parses values from JSON files and includes them in defines list', () async {
+        fileSystem.file(fileSystem.path.join('lib', 'main.dart')).createSync(recursive: true);
+        fileSystem.file('pubspec.yaml').createSync();
+        fileSystem.file('.packages').createSync();
+        await fileSystem.file('config1.json').writeAsString(
+          '''
+            {
+              "kInt": 1,
+              "kDouble": 1.1,
+              "name": "denghaizhu",
+              "title": "this is title from config json file",
+              "nullValue": null,
+              "containEqual": "sfadsfv=432f"
+            }
+          '''
+        );
+        await fileSystem.file('config2.json').writeAsString(
+            '''
+            {
+              "body": "this is body from config json file"
+            }
+          '''
+        );
+
+        await dummyCommandRunner.run(<String>[
+          'dummy',
+          '--dart-define-from-file=config1.json',
+          '--dart-define-from-file=config2.json',
+        ]);
+
+        final BuildInfo buildInfo = await dummyCommand.getBuildInfo(forcedBuildMode: BuildMode.debug);
+        expect(buildInfo.dartDefines, containsAll(const <String>[
+          'kInt=1',
+          'kDouble=1.1',
+          'name=denghaizhu',
+          'title=this is title from config json file',
+          'nullValue=null',
+          'containEqual=sfadsfv=432f',
+          'body=this is body from config json file',
+        ]));
+      }, overrides: <Type, Generator>{
+        FileSystem: () => fileSystem,
+        Logger: () => logger,
+        FileSystemUtils: () => fileSystemUtils,
+        Platform: () => platform,
+        ProcessManager: () => processManager,
+      });
+
+      testUsingContext('has values with identical keys from --dart-define take precedence', () async {
+        fileSystem
+          .file(fileSystem.path.join('lib', 'main.dart'))
+          .createSync(recursive: true);
+        fileSystem.file('pubspec.yaml').createSync();
+        fileSystem.file('.packages').createSync();
+        fileSystem.file('.env').writeAsStringSync('''
+            MY_VALUE=VALUE_FROM_ENV_FILE
+          ''');
+
+        await dummyCommandRunner.run(<String>[
+          'dummy',
+          '--dart-define=MY_VALUE=VALUE_FROM_COMMAND',
+          '--dart-define-from-file=.env',
+        ]);
+
+        final BuildInfo buildInfo = await dummyCommand.getBuildInfo(forcedBuildMode: BuildMode.debug);
+        expect(buildInfo.dartDefines, containsAll(const <String>[
+          'MY_VALUE=VALUE_FROM_ENV_FILE',
+          'MY_VALUE=VALUE_FROM_COMMAND',
+        ]));
+      }, overrides: <Type, Generator>{
+        FileSystem: () => fileSystem,
+        Logger: () => logger,
+        FileSystemUtils: () => fileSystemUtils,
+        Platform: () => platform,
+        ProcessManager: () => processManager,
+      });
+
+      testUsingContext('correctly parses a valid env file', () async {
+        fileSystem
+            .file(fileSystem.path.join('lib', 'main.dart'))
+            .createSync(recursive: true);
+        fileSystem.file('pubspec.yaml').createSync();
+        fileSystem.file('.packages').createSync();
+        await fileSystem.file('.env').writeAsString('''
+            # comment
+            kInt=1
+            kDouble=1.1 # should be double
+
+            name=piotrfleury
+            title=this is title from config env file
+            empty=
+
+            doubleQuotes="double quotes 'value'#=" # double quotes
+            singleQuotes='single quotes "value"#=' # single quotes
+            backQuotes=`back quotes "value" '#=` # back quotes
+
+            hashString="some-#-hash-string-value"
+
+            # Play around with spaces around the equals sign.
+            spaceBeforeEqual =value
+            spaceAroundEqual = value
+            spaceAfterEqual= value
+
+          ''');
+        await fileSystem.file('.env2').writeAsString('''
+            # second comment
+
+            body=this is body from config env file
+          ''');
+
+        await dummyCommandRunner.run(<String>[
+          'dummy',
+          '--dart-define-from-file=.env',
+          '--dart-define-from-file=.env2',
+        ]);
+
+        final BuildInfo buildInfo = await dummyCommand.getBuildInfo(forcedBuildMode: BuildMode.debug);
+        expect(buildInfo.dartDefines, containsAll(const <String>[
+          'kInt=1',
+          'kDouble=1.1',
+          'name=piotrfleury',
+          'title=this is title from config env file',
+          'empty=',
+          "doubleQuotes=double quotes 'value'#=",
+          'singleQuotes=single quotes "value"#=',
+          'backQuotes=back quotes "value" \'#=',
+          'hashString=some-#-hash-string-value',
+          'spaceBeforeEqual=value',
+          'spaceAroundEqual=value',
+          'spaceAfterEqual=value',
+          'body=this is body from config env file'
+        ]));
+      }, overrides: <Type, Generator>{
+        FileSystem: () => fileSystem,
+        Logger: () => logger,
+        FileSystemUtils: () => fileSystemUtils,
+        Platform: () => platform,
+        ProcessManager: () => processManager,
+      });
+
+      testUsingContext('throws a ToolExit when .env file contains a multiline value', () async {
+        fileSystem
+            .file(fileSystem.path.join('lib', 'main.dart'))
+            .createSync(recursive: true);
+        fileSystem.file('pubspec.yaml').createSync();
+        fileSystem.file('.packages').createSync();
+        await fileSystem.file('.env').writeAsString('''
+            # single line value
+            name=piotrfleury
+
+            # multi-line value
+            multiline = """ Welcome to .env demo
+            a simple counter app with .env file support
+            for more info, check out the README.md file
+            Thanks! """ # This is the welcome message that will be displayed on the counter app
+
+          ''');
+
+        await dummyCommandRunner.run(<String>[
+          'dummy',
+          '--dart-define-from-file=.env',
+        ]);
+        expect(dummyCommand.getBuildInfo(forcedBuildMode: BuildMode.debug),
+          throwsToolExit(message: 'Multi-line value is not supported: multiline = """ Welcome to .env demo'));
+      }, overrides: <Type, Generator>{
+        FileSystem: () => fileSystem,
+        Logger: () => logger,
+        FileSystemUtils: () => fileSystemUtils,
+        Platform: () => platform,
+        ProcessManager: () => processManager,
+      });
+
+      testUsingContext('works with mixed file formats',
+          () async {
+        fileSystem
+            .file(fileSystem.path.join('lib', 'main.dart'))
+            .createSync(recursive: true);
+        fileSystem.file('pubspec.yaml').createSync();
+        fileSystem.file('.packages').createSync();
+        await fileSystem.file('.env').writeAsString('''
+            kInt=1
+            kDouble=1.1
+            name=piotrfleury
+            title=this is title from config env file
+          ''');
+        await fileSystem.file('config.json').writeAsString('''
+            {
+              "body": "this is body from config json file"
+            }
+          ''');
+
+        await dummyCommandRunner.run(<String>[
+          'dummy',
+          '--dart-define-from-file=.env',
+          '--dart-define-from-file=config.json',
+        ]);
+
+        final BuildInfo buildInfo = await dummyCommand.getBuildInfo(forcedBuildMode: BuildMode.debug);
+        expect(buildInfo.dartDefines, containsAll(const <String>[
+          'kInt=1',
+          'kDouble=1.1',
+          'name=piotrfleury',
+          'title=this is title from config env file',
+          'body=this is body from config json file',
+        ]));
+      }, overrides: <Type, Generator>{
+        FileSystem: () => fileSystem,
+        Logger: () => logger,
+        FileSystemUtils: () => fileSystemUtils,
+        Platform: () => platform,
+        ProcessManager: () => processManager,
+      });
+
+      testUsingContext('when files contain entries with duplicate keys, uses the value from the lattermost file', () async {
+        fileSystem.file(fileSystem.path.join('lib', 'main.dart')).createSync(recursive: true);
+        fileSystem.file('pubspec.yaml').createSync();
+        fileSystem.file('.packages').createSync();
+        await fileSystem.file('config1.json').writeAsString(
+            '''
+            {
+              "kInt": 1,
+              "kDouble": 1.1,
+              "name": "denghaizhu",
+              "title": "this is title from config json file"
+            }
+          '''
+        );
+        await fileSystem.file('config2.json').writeAsString(
+            '''
+            {
+              "kInt": "2"
+            }
+          '''
+        );
+
+        await dummyCommandRunner.run(<String>[
+          'dummy',
+          '--dart-define-from-file=config1.json',
+          '--dart-define-from-file=config2.json',
+        ]);
+        final BuildInfo buildInfo = await dummyCommand.getBuildInfo(forcedBuildMode: BuildMode.debug);
+        expect(buildInfo.dartDefines, containsAll(const <String>[
+          'kInt=2',
+          'kDouble=1.1',
+          'name=denghaizhu',
+          'title=this is title from config json file'
+        ]));
+      }, overrides: <Type, Generator>{
+        FileSystem: () => fileSystem,
+        Logger: () => logger,
+        FileSystemUtils: () => fileSystemUtils,
+        Platform: () => platform,
+        ProcessManager: () => processManager,
+      });
+
+      testUsingContext('throws a ToolExit when the file cannot be parsed', () async {
+        fileSystem.file(fileSystem.path.join('lib', 'main.dart')).createSync(recursive: true);
+        fileSystem.file('pubspec.yaml').createSync();
+        fileSystem.file('.packages').createSync();
+        fileSystem.directory('config').createSync();
+
+        await dummyCommandRunner.run(<String>[
+          'dummy',
+          '--dart-define-from-file=config',
+        ]);
+        expect(dummyCommand.getBuildInfo(forcedBuildMode: BuildMode.debug),
+          throwsToolExit(message: 'Did not find the file passed to "--dart-define-from-file". Path: config'));
+      }, overrides: <Type, Generator>{
+        FileSystem: () => fileSystem,
+        Logger: () => logger,
+        FileSystemUtils: () => fileSystemUtils,
+        Platform: () => platform,
+        ProcessManager: () => processManager,
+      });
+
+      testUsingContext('throws a ToolExit when the given JSON file is malformed', () async {
+        fileSystem.file(fileSystem.path.join('lib', 'main.dart')).createSync(recursive: true);
+        fileSystem.file('pubspec.yaml').createSync();
+        fileSystem.file('.packages').createSync();
+        await fileSystem.file('config.json').writeAsString(
+          '''
+            {
+              "kInt": 1Error json format
+              "kDouble": 1.1,
+              "name": "denghaizhu",
+              "title": "this is title from config json file"
+            }
+          '''
+        );
+
+        await dummyCommandRunner.run(<String>[
+          'dummy',
+          '--dart-define-from-file=config.json',
+        ]);
+        expect(dummyCommand.getBuildInfo(forcedBuildMode: BuildMode.debug),
+          throwsToolExit(message: 'Json config define file "--dart-define-from-file=config.json" format err'));
+      }, overrides: <Type, Generator>{
+        FileSystem: () => fileSystem,
+        Logger: () => logger,
+        FileSystemUtils: () => fileSystemUtils,
+        Platform: () => platform,
+        ProcessManager: () => processManager,
+      });
+
+      testUsingContext('test --dart-define-from-file option with err file format', () async {
+        fileSystem.directory('config').createSync();
+        await dummyCommandRunner.run(<String>[
+          'dummy',
+          '--dart-define=k=v',
+          '--dart-define-from-file=config']);
+        expect(dummyCommand.getBuildInfo(forcedBuildMode: BuildMode.debug),
+            throwsToolExit(message: 'Did not find the file passed to "--dart-define-from-file". Path: config'));
+      }, overrides: <Type, Generator>{
+        FileSystem: () => fileSystem,
+        Logger: () => logger,
+        FileSystemUtils: () => fileSystemUtils,
+        Platform: () => platform,
+        ProcessManager: () => processManager,
+      });
+
+      testUsingContext('test --dart-define-from-file option with err json format', () async {
+        await fileSystem.file('config.json').writeAsString(
+            '''
+              {
+                "kInt": 1,
+                "kDouble": 1.1,
+                "name": "err json format,
+                "title": "this is title from config json file"
+              }
+            '''
+        );
+
+        await dummyCommandRunner.run(<String>[
+          'dummy',
+          '--dart-define=k=v',
+          '--dart-define-from-file=config.json']);
+
+        expect(dummyCommand.getBuildInfo(forcedBuildMode: BuildMode.debug),
+            throwsToolExit(message: 'Json config define file "--dart-define-from-file=config.json" format err, please fix first! format err:'));
+      }, overrides: <Type, Generator>{
+        FileSystem: () => fileSystem,
+        Logger: () => logger,
+        FileSystemUtils: () => fileSystemUtils,
+        Platform: () => platform,
+        ProcessManager: () => processManager,
+      });
+    });
+
     group('--flavor', () {
       late _TestDeviceManager testDeviceManager;
       late Logger logger;
@@ -935,4 +1302,14 @@ class _TestRunCommandThatOnlyValidates extends RunCommand {
   Future<FlutterCommandResult> runCommand() async {
     return FlutterCommandResult.success();
   }
+}
+
+
+Iterable<String> _decodeDartDefines(Environment environment) {
+  final String encodedDefines = environment.defines[kDartDefines]!;
+  const Utf8Decoder byteDecoder = Utf8Decoder();
+  return encodedDefines
+      .split(',')
+      .map<Uint8List>(base64.decode)
+      .map<String>(byteDecoder.convert);
 }
